+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Allow preview current edits on [indoor=](https://indoorequal.org/) (thanks to François de Metz)
 
+### Fixed
+- Prevent empty tag key or value to be sent to OSM
+
 
 ## 1.0.10 - 2021-08-16